fix the action_status command
action_runner produces lines of yaml, whilst this script assumed that the
status file is a single blob of yaml. fix action_status.py to

1) parse each line as yaml individually
2) take the last line as the source of truth for any fields to be
printed

+++ bin/action_status.py
@@ -3,9 +3,9 @@
 Read values from a status file created by action_runner.py
 """
 from __future__ import absolute_import
+from __future__ import print_function
 from __future__ import unicode_literals
 
-import functools
 import logging
 import os
 import signal
@@ -13,42 +13,39 @@ import sys
 
 from tron import yaml
 
-
 log = logging.getLogger('tron.action_status')
-
 
 STATUS_FILE = 'status'
 
 
-def print_field(field, status_file):
-    sys.stdout.write(str(status_file[field]))
+def get_field(field, status_file):
+    lines = status_file.readlines()
+    content = yaml.load(lines[-1])
+    return content.get(field)
 
 
 def print_status_file(status_file):
-    yaml.dump(status_file, sys.stdout)
+    for line in status_file.readlines():
+        print(yaml.load(line))
 
 
 def send_signal(signal_num, status_file):
-    pid = status_file['pid']
-    try:
-        os.kill(pid, signal_num)
-    except OSError as e:
-        msg = "Failed to signal %s with %s: %s"
-        raise SystemExit(msg % (pid, signal_num, e))
+    pid = get_field('pid', status_file)
+    if pid:
+        try:
+            os.kill(pid, signal_num)
+        except OSError as e:
+            msg = "Failed to signal %s with %s: %s"
+            raise SystemExit(msg % (pid, signal_num, e))
 
 
 commands = {
     'print':        print_status_file,
-    'pid':          functools.partial(print_field, 'pid'),
-    'return_code':  functools.partial(print_field, 'return_code'),
-    'terminate':    functools.partial(send_signal, signal.SIGTERM),
-    'kill':         functools.partial(send_signal, signal.SIGKILL),
+    'pid': lambda statusfile: print(get_field('pid', statusfile)),
+    'return_code': lambda statusfile: print(get_field('return_code', statusfile)),
+    'terminate': lambda statusfile: send_signal(signal.SIGTERM. statusfile),
+    'kill': lambda statusfile: send_signal(signal.SIGKILL, statusfile),
 }
-
-
-def get_status_file(path):
-    with open(path, 'r') as fh:
-        return yaml.load(fh)
 
 
 def parse_args(args):
@@ -68,5 +65,5 @@ def run_command(command, status_file):
 if __name__ == "__main__":
     logging.basicConfig()
     path, command = parse_args(sys.argv)
-    status_file = get_status_file(os.path.join(path, STATUS_FILE))
-    run_command(command, status_file)
+    with open(os.path.join(path, STATUS_FILE)) as f:
+        run_command(command, f)
